REFACTOR: functions inside wheelchair_utils folder
