+++ ocp_propulsion.py
@@ -4,6 +4,7 @@ It needs to be a example that simulates the pushing phase of a propulsion cycle 
 i.e. with the close-loop on the handrim, and the rolling joint constraint.
 We need to get inspiration from the work of Anais.
 """
+
 import numpy as np
 from biorbd_casadi import marker_index
 from casadi import MX, SX, vertcat, Function, jacobian, sqrt, atan2
@@ -28,16 +29,16 @@ from bioptim import (
     PhaseDynamics,
     HolonomicConstraintsList,
 )
-from custom_biorbd_model_holonomic import BiorbdModelCustomHolonomic
+from wheelchair_utils.custom_biorbd_model_holonomic import BiorbdModelCustomHolonomic
 
 
 def custom_dynamic(
-        time: MX | SX,
-        states: MX | SX,
-        controls: MX | SX,
-        parameters: MX | SX,
-        stochastic_variables: MX | SX,
-        nlp: NonLinearProgram,
+    time: MX | SX,
+    states: MX | SX,
+    controls: MX | SX,
+    parameters: MX | SX,
+    stochastic_variables: MX | SX,
+    nlp: NonLinearProgram,
 ) -> DynamicsEvaluation:
     """
     The custom dynamics function that provides the derivative of the states: dxdt = f(x, u, p)
@@ -81,25 +82,25 @@ def custom_configure(ocp: OptimalControlProgram, nlp: NonLinearProgram):
 
     name_u = [nlp.model.name_dof[i] for i in range(nlp.model.nb_independent_joints)]
     axes_idx = ConfigureProblem._apply_phase_mapping(ocp, nlp, "u")
-    ConfigureProblem.configure_new_variable(
-        "u", name_u, ocp, nlp, True, False, False, axes_idx=axes_idx
-    )
+    ConfigureProblem.configure_new_variable("u", name_u, ocp, nlp, True, False, False, axes_idx=axes_idx)
 
     name = "udot"
     name_qdot = ConfigureProblem._get_kinematics_based_names(nlp, "qdot")
     name_udot = [name_qdot[i] for i in range(nlp.model.nb_independent_joints)]
     axes_idx = ConfigureProblem._apply_phase_mapping(ocp, nlp, name)
-    ConfigureProblem.configure_new_variable(
-        name, name_udot, ocp, nlp, True, False, False, axes_idx=axes_idx
-    )
+    ConfigureProblem.configure_new_variable(name, name_udot, ocp, nlp, True, False, False, axes_idx=axes_idx)
 
     ConfigureProblem.configure_tau(ocp, nlp, as_states=False, as_controls=True)
     ConfigureProblem.configure_dynamics_function(ocp, nlp, custom_dynamic)
 
 
 def generate_close_loop_constraint(
-        biorbd_model, marker_1: str, marker_2: str, index: slice = slice(0, 3), local_frame_index: int = None,
-        parameters: MX = MX(),
+    biorbd_model,
+    marker_1: str,
+    marker_2: str,
+    index: slice = slice(0, 3),
+    local_frame_index: int = None,
+    parameters: MX = MX(),
 ) -> tuple[Function, Function, Function]:
     """Generate a close loop constraint between two markers"""
 
@@ -141,7 +142,7 @@ def generate_close_loop_constraint(
 
     # the double derivative of the constraint
     constraint_double_derivative = (
-            constraint_jacobian_func(q_sym) @ q_ddot_sym + constraint_jacobian_func(q_dot_sym) @ q_dot_sym
+        constraint_jacobian_func(q_sym) @ q_ddot_sym + constraint_jacobian_func(q_dot_sym) @ q_dot_sym
     )
 
     constraint_double_derivative_func = Function(
@@ -156,11 +157,11 @@ def generate_close_loop_constraint(
 
 
 def generate_close_loop_constraint_polar_coordinates(
-        biorbd_model: BiorbdModelCustomHolonomic,
-        marker_1: str,
-        wheel_frame_index: int = None,
-        handrim_radius: float = 0.35,
-        contact_angle: float = None,
+    biorbd_model: BiorbdModelCustomHolonomic,
+    marker_1: str,
+    wheel_frame_index: int = None,
+    handrim_radius: float = 0.35,
+    contact_angle: float = None,
 ) -> tuple[Function, Function, Function]:
     """
     Generate a close loop constraint between two markers, in polar coordinates for the wheel.
@@ -223,8 +224,8 @@ def generate_close_loop_constraint_polar_coordinates(
 
     # the double derivative of the constraint
     constraint_double_derivative = (
-            constraint_jacobian_func(q_sym) @ q_ddot_sym + jacobian(constraint_jacobian_func(q_sym) @ q_dot_sym,
-                                                                    q_sym) @ q_dot_sym
+        constraint_jacobian_func(q_sym) @ q_ddot_sym
+        + jacobian(constraint_jacobian_func(q_sym) @ q_dot_sym, q_sym) @ q_dot_sym
     )
 
     constraint_double_derivative_func = Function(
@@ -239,10 +240,10 @@ def generate_close_loop_constraint_polar_coordinates(
 
 
 def generate_rolling_joint_constraint(
-        biorbd_model: BiorbdModelCustomHolonomic,
-        translation_joint_index: int,
-        rotation_joint_index: int,
-        radius: float = 1,
+    biorbd_model: BiorbdModelCustomHolonomic,
+    translation_joint_index: int,
+    rotation_joint_index: int,
+    radius: float = 1,
 ) -> tuple[Function, Function, Function]:
     """Generate a rolling joint constraint between two joints"""
 
@@ -272,8 +273,8 @@ def generate_rolling_joint_constraint(
     ).expand()
 
     constraint_double_derivative = (
-            constraint_jacobian_func(q_sym) @ q_ddot_sym + jacobian(constraint_jacobian_func(q_sym) @ q_dot_sym,
-                                                                    q_sym) @ q_dot_sym
+        constraint_jacobian_func(q_sym) @ q_ddot_sym
+        + jacobian(constraint_jacobian_func(q_sym) @ q_dot_sym, q_sym) @ q_dot_sym
     )
 
     constraint_double_derivative_func = Function(
@@ -288,9 +289,9 @@ def generate_rolling_joint_constraint(
 
 
 def prepare_ocp(
-        biorbd_model_path: str,
-        ode_solver: OdeSolverBase = OdeSolver.RK4(),
-        n_shooting=50,
+    biorbd_model_path: str,
+    ode_solver: OdeSolverBase = OdeSolver.RK4(),
+    n_shooting=50,
 ) -> OptimalControlProgram:
     """
     Prepare the program
@@ -323,10 +324,9 @@ def prepare_ocp(
         radius=0.35,
     )
 
-    bio_model.set_holonomic_configuration(constraints_list=holonomic_constraints,
-                                          independent_joint_index=[0, 2, 3],
-                                          dependent_joint_index=[1]
-                                          )
+    bio_model.set_holonomic_configuration(
+        constraints_list=holonomic_constraints, independent_joint_index=[0, 2, 3], dependent_joint_index=[1]
+    )
 
     # constraint, constraint_jacobian, constraint_double_derivative = generate_close_loop_constraint(
     #     biorbd_model=bio_model,
@@ -361,8 +361,9 @@ def prepare_ocp(
 
     # Dynamics
     dynamics = DynamicsList()
-    dynamics.add(custom_configure, dynamic_function=custom_dynamic,
-                 phase_dynamics=PhaseDynamics.SHARED_DURING_THE_PHASE)
+    dynamics.add(
+        custom_configure, dynamic_function=custom_dynamic, phase_dynamics=PhaseDynamics.SHARED_DURING_THE_PHASE
+    )
 
     # Path Constraints
     constraints = ConstraintList()
@@ -392,23 +393,26 @@ def prepare_ocp(
     u_init.add("tau", [tau_init] * 2)
     # ------------- #
 
-    return OptimalControlProgram(
+    return (
+        OptimalControlProgram(
+            bio_model,
+            dynamics=dynamics,
+            n_shooting=n_shooting,
+            phase_time=final_time,
+            x_init=x_init,
+            u_init=u_init,
+            x_bounds=x_bounds,
+            u_bounds=u_bounds,
+            objective_functions=objective_functions,
+            constraints=constraints,
+            ode_solver=ode_solver,
+            use_sx=False,
+            variable_mappings=variable_bimapping,
+            parameters=parameters,
+            n_threads=8,
+        ),
         bio_model,
-        dynamics=dynamics,
-        n_shooting=n_shooting,
-        phase_time=final_time,
-        x_init=x_init,
-        u_init=u_init,
-        x_bounds=x_bounds,
-        u_bounds=u_bounds,
-        objective_functions=objective_functions,
-        constraints=constraints,
-        ode_solver=ode_solver,
-        use_sx=False,
-        variable_mappings=variable_bimapping,
-        parameters=parameters,
-        n_threads=8,
-    ), bio_model
+    )
 
 
 def main():
@@ -432,6 +436,7 @@ def main():
         q[:, i] = qi
 
     import bioviz
+
     viz = bioviz.Viz(model_path)
     viz.load_movement(q)
     viz.exec()
