+++ wheelchair_utils/__init__.py
@@ -1,0 +1,3 @@
+from .custom_biorbd_model_holonomic import BiorbdModelCustomHolonomic
+from .dynamics import custom_dynamic, custom_configure, compute_all_states
+from .holon_constraints import generate_close_loop_constraint, generate_rolling_joint_constraint
