+++ wheelchair_utils/custom_biorbd_model_holonomic.py
